LayoutTests/webaudio/WaveShaper/waveshaper-nan-infinity-handling-expected.txt

Tests WaveShaperDSPKernel::processCurveWithData handling of NaN, Infinity, and normal input values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS isNaN(output[0]) is false
PASS isFinite(output[0]) is true
PASS isNaN(output[1]) is false
PASS isFinite(output[1]) is true
PASS isNaN(output[2]) is false
PASS isFinite(output[2]) is true
PASS isNaN(output[3]) is false
PASS isFinite(output[3]) is true
PASS 0 is approximately 0
PASS 0 is approximately 0
PASS 0 is approximately 0
PASS 0.5 is approximately 0.5
PASS 1 is approximately 1
PASS 0.5 is approximately 0.5
PASS isFinite(output[1]) is true
PASS 0.75 is approximately 0.75
PASS isFinite(output[3]) is true
PASS 0 is approximately 0
PASS 0.5 is approximately 0.5
PASS successfullyParsed is true

TEST COMPLETE
